+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Betty Cropper Change Log
 
+## Version 2.5.3
+
+- Fix `Image.clear_crops()` to handle when save crops path is different than image source path. Previously was preventing saved-to-disk crops from being deleted.
+
+## Version 2.5.2
+
+- Image source caching attempts to use named cache "storage", else "default".
+
 ## Version 2.5.0
 
 - Add storage caching layer (ex: memcached in front of S3 backend) to improve performance with slower storage backends.